Moved store creation to separate file, removed extra calcuclateWinner function, reorganized some files.
