+++ tic-tac-toe/src/app/components/App/index.js
@@ -1,25 +1,12 @@
-import React from 'react';
+import React, { Component }from 'react';
 import { Provider } from 'react-redux';
-import { createStore } from 'redux';
 
-import { handleClick as gameReducer } from '~redux/game/reducers.js';
+import store from '~redux/store';
 
 import Game from '../Game';
 
-const initState = {
-  history: [
-    {
-      squares: Array(9).fill(null)
-    }
-  ],
-  xIsNext: true,
-  stepNumber: 0,
-  winner: null
-};
 
-const store = createStore(gameReducer, initState);
-
-class App extends React.Component {
+class App extends Component {
   render() {
     return (
       <Provider store={store}>
