+++ tic-tac-toe/src/redux/game/reducers.js
@@ -1,15 +1,5 @@
-function calculateWinner(squares) {
-  const lines = [[0, 1, 2], [3, 4, 5], [6, 7, 8], [0, 3, 6], [1, 4, 7], [2, 5, 8], [0, 4, 8], [2, 4, 6]];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
-}
+import calculateWinner from '~utils/gameUtils';
 
-// REDUCERS
 export function handleClick(state, action) {
   const history = state.history.slice(0, state.stepNumber + 1);
   const current = history[history.length - 1];
